redo parsers

+++ src/term.rs
@@ -4,7 +4,7 @@ use std::collections::HashMap;
 pub enum VTerm {
     Var { name: String },
     Thunk { t: Box<CTerm> },
-    Int { value: i32 },
+    Int { value: i64 },
     Str { value: String },
     Tuple { values: Vec<VTerm> },
 }
@@ -16,7 +16,7 @@ pub enum CTerm {
     Force { thunk: VTerm },
     Let { t: Box<CTerm>, bound_name: String, body: Box<CTerm> },
     Def { name: String },
-    CaseInt { t: VTerm, branches: HashMap<i32, CTerm>, default_branch: Option<Box<CTerm>> },
+    CaseInt { t: VTerm, branches: HashMap<i64, CTerm>, default_branch: Option<Box<CTerm>> },
     CaseTuple { t: VTerm, bound_names: Vec<String>, branch: Box<CTerm> },
     CaseStr { t: VTerm, branches: HashMap<String, CTerm>, default_branch: Option<Box<CTerm>> },
     Primitive { name: &'static str, arity: u8 },
